# GeekShopping.CartAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GeekShopping.CartAPI",
    "version": "v1"
  },
  "paths": {
    "/find-cart/{id}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        }
      }
    },
    "/add-cart": {
      "post": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        }
      }
    },
    "/update-cart": {
      "put": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        }
      }
    },
    "/remove-cart/{id}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        }
      }
    },
    "/apply-coupon": {
      "post": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        }
      }
    },
    "/remove-coupon/{userId}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CartVO"
            }
          }
        }
      }
    },
    "/checkout": {
      "post": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CheckoutHeaderVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CheckoutHeaderVO"
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CartDetailVO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "cartHeaderId": {
          "format": "int64",
          "type": "integer"
        },
        "cartHeader": {
          "$ref": "#/definitions/CartHeaderVO"
        },
        "productId": {
          "format": "int64",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/ProductVO"
        },
        "count": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CartHeaderVO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "userId": {
          "type": "string"
        },
        "couponCode": {
          "type": "string"
        }
      }
    },
    "CartVO": {
      "type": "object",
      "properties": {
        "cartHeader": {
          "$ref": "#/definitions/CartHeaderVO"
        },
        "cartDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartDetailVO"
          }
        }
      }
    },
    "CheckoutHeaderVO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "messageCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "couponCode": {
          "type": "string"
        },
        "purchaseAmount": {
          "format": "double",
          "type": "number"
        },
        "discountAmount": {
          "format": "double",
          "type": "number"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "dateTime": {
          "format": "date-time",
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "cardNumber": {
          "type": "string"
        },
        "cvv": {
          "type": "string"
        },
        "expiryMothYear": {
          "type": "string"
        },
        "cartTotalItens": {
          "format": "int32",
          "type": "integer"
        },
        "cartDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartDetailVO"
          }
        }
      }
    },
    "ProductVO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "description": {
          "type": "string"
        },
        "categoryName": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        }
      }
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter 'Bearer' [space] and your token!"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}